On 32-bit x86, get the vsyscall address from the AUX vector
On 32-bit x86, get the vsyscall address from the AT_SYSINFO AUX vector
entry, rather than looking it up in the vDSO. This avoids the need to
link in all the vDSO code if it isn't otherwise needed. And, it's
simpler, avoiding the need for the `rustix_int_0x80` function.